Separate web requests, celery tasks, and run_all scripts into separate modules for each project
