+++ otakaro/pollution_model/pollution_tables.py
@@ -4,6 +4,7 @@ from datetime import datetime, timezone
 
 from geoalchemy2 import Geometry
 from sqlalchemy import Column, Integer, String, Float, DateTime
+from sqlalchemy.engine import Engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import DeclarativeMeta
 
@@ -197,3 +198,28 @@ class MedusaScenarios(Base):
 
     created_at = Column(DateTime(timezone=True), default=datetime.now(timezone.utc), comment="log created datetime")
     geometry = Column(Geometry("POLYGON", srid=2193))
+
+
+def define_custom_db_functions(engine: Engine) -> None:
+    """
+    Create custom SQL functions for the database, to aid in queries.
+
+    Parameters
+    ----------
+    engine : Engine
+        The engine used to connect to the database.
+    """
+    # Create function for calculating result as significant figures
+    create_significant_figures_statement = """
+        CREATE OR REPLACE FUNCTION sig_fig(n double precision, digits integer)
+            RETURNS numeric
+        AS
+        $$
+        SELECT CASE
+                   WHEN n = 0 THEN 0
+                   ELSE round(n::numeric, digits - 1 - floor(log(abs(n)))::int)
+                   END
+        $$ LANGUAGE sql IMMUTABLE
+                        STRICT;
+    """
+    engine.execute(create_significant_figures_statement)
